requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
cryptography>=3.1
scikit-learn>=1.3.0
networkx>=3.2.0
numpy>=1.24.0
pytest>=7.0.0
